refactor: Rearrange package structure
