+++ src/main/java/com/woowacourse/lotto/domain/Lotto.java
@@ -1,5 +1,7 @@
 package com.woowacourse.lotto.domain;
 
+import com.woowacourse.lotto.util.LottoGenerator;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
