+++ src/main/java/com/woowacourse/lotto/util/LottoGenerator.java
@@ -1,4 +1,4 @@
-package com.woowacourse.lotto.domain;
+package com.woowacourse.lotto.util;
 
 public interface LottoGenerator {
     int generateNumber();
