+++ src/test/java/com/woowacourse/lotto/domain/RandomLottoGeneratorTest.java
@@ -1,5 +1,7 @@
 package com.woowacourse.lotto.domain;
 
+import com.woowacourse.lotto.util.LottoGenerator;
+import com.woowacourse.lotto.util.RandomLottoGenerator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
